More clang-tidy fixes.

+++ src/KafkaW/Producer.cpp
@@ -86,7 +86,7 @@ Producer::~Producer() {
   LOG(Sev::Debug, "~Producer");
   if (RdKafkaPtr) {
     int TimeoutMS = 1;
-    uint32_t OutQueueLength = 0;
+    int OutQueueLength = 0;
     while (true) {
       OutQueueLength = rd_kafka_outq_len(RdKafkaPtr);
       if (OutQueueLength == 0) {
@@ -175,7 +175,7 @@ void Producer::poll() {
 
 rd_kafka_t *Producer::getRdKafkaPtr() const { return RdKafkaPtr; }
 
-uint64_t Producer::outputQueueLength() { return rd_kafka_outq_len(RdKafkaPtr); }
+int Producer::outputQueueLength() { return rd_kafka_outq_len(RdKafkaPtr); }
 
 ProducerStats::ProducerStats(ProducerStats const &x) {
   produced = x.produced.load();
